[When] Update when temp value scope (#1157)
* [mlir] move finalize when pass to end of sequence

Then when finalize must occur after any passes that modify the circuit

e.g. `insert_coreir_wires` could flatten a value used in the
conditional logic, which then requires updating the when builder
instance

* Add build dir

* Skip buggy test

* Move test, update comment

* [when] update test/impl for when temp value scope

Updates the test so the temporary value is only referenced inside the
scope it's created.

Updates the insert_coreir_wires logic to handle builders (needed for the
when builder which hasn't been finalized).

I think this is probably a better pattern than allowing users to abuse
the fact that the Python namespace allows values to fall through to the
enclosing scope, otherwise we run into potential issues with latch
inference (either we need to require the user to assign the value in all
when contexts, or determine when a value is referenced outside it's
context and change its enclosing scope).  While possible, I doubt
there's many useful patterns that motivate this added complexity.

* Update gold

* Add comment

* Update magma/when.py

Co-authored-by: rsetaluri <rsetaluri@users.noreply.github.com>

Co-authored-by: rsetaluri <rsetaluri@users.noreply.github.com>