[CompileGuard] Support simple anon drivees
For now, we only support anon values that are driving values with the
same status: either all must be external or internal. We also assume
that more drivees will not be added after the fact. We still do not
support unwired anon values, nor do we support anon values that have
wirings that change after the compile_guard (e.g. an internal value is
then wired to an external value after the compile guard)

In the future, we could improve support by:
(1) staging the resolution of externality until the end of definition
(2) copying anon values that drive mixed values (one copy for
    external, one for internal)
(3) doing final wiring for internal/external copies with the proper
    values