[Generator2] Improve class creation pipeline
Currently we have a guard for when the Generator2 class is created. In
fact, we need a more general check for when any class is created
has the metaclass Generator2Meta.

One hacky way to determine that this is a type(...) call, i.e. it is
invoked in the class creation pipeline. In that case, there are exactly
3 arguments of the form name, bases, dct. In theory, it would be
possible for a Generator subclass to have the same argument number and
types, but having the "__module__" member of the dct would be highly
unlikely. Furthmore they should be using a hashable dict, so it
shouldn't satisfy that check. Note that this is why we use "type is"
instead of isinstance.