[CompileGuard] Avoid name collision in port logic
Since a value's name is not guaranteed to be unique (e.g. inline wire's
get `I` for the wire instance port name), there's a possibility for a
name collision.  Since the compile_guard logic is hidden to the user, I
think for simplicity's sake it's easiest to just avoid the collision by
not giving the port an explicit name in the clock case.