[When] Allow user to override implicit logic
In the spirit of explicit is better than implicit, I think it's best if
we don't require the user to use `rewire` if they want to override the
implicit wire logic.  This changes the logic so that if we encounter a
wire called on something implicitly wired, we remove the implicit wire
(without raising an already driven error).  For this to work, we need to
update the when builder code to mark that it is wiring implicitly when
adding the drivee port logic.