[When] Fix bug in resolve logic for when output (#1187)
* [When] Fix bug in resolve logic for when output

If the user call's .value() on a conditionally driven value, they may
leave a hanging reference to a when output that was resolved.  When a
value is resolved, it's children are added as output ports to the
builder and original output "discarded".  To handle these stale
references, at finalization we can traverse the outputs and update
anything still driven by a stale output with the resolved children.