[MLIR] Attach location info for magma instances
All ops derived from magma instances will be annotated with the location
info from the containing magma instance.

Note that this may not always be "correct". For example, the assignment
op for a register will be annotated with the location from the register
instantiation, when we would like it to come from the wiring statement.