Delay Bulk Resolve (#1184)
The essence of this change is that we delay the resolution of bulk wires to when children are wired. So, referencing an index of an array doesn't trigger bulk resolve until that child is wired. This allows us to maintain bulk wires in more places, even though, for example, we might use a single bit of an array to drive a conversion, we don't need to walk back and resolve everything related to the container array.

To do this, we need to introduce a parent reference for values that are derived from an array/tuple. Related, I changed the view parent field to be parent_view to avoid a name collision there.

I also updated the backend logic to expect either children or slices when doing an array create.