[when] Fix bug in ctx inheritance and builder ctx
Tuple.set_enclosing_when_context was not updating the current value
before recursing, also updated the logic to only recurse on elaborated
value (avoiding elaboration if possible, future values will inherit the
set context).

Adding a port to the when builder primitive should be in no context
which is necessary for conditional wiring checks (builder wiring should
be unconditional).