[When] Use concat/create logic for elaborated arrays (#1228)
Improves the when logic to reuse child array elements for output driver ports, which allows us to maintain a single array assignment statement inside the when statement, even if it's wired up by children. The backend logic is updated to handle this by inserting a concat or create operation for the array assignment. We no longer need the update resolved ref logic since the when builder outputs do not introduce new ports for elaborated elements.

NOTE I added a test case that is commented out that demonstrates a case where we could "pack" assignments to reduce elaboration in the when statements, however this is a different (yet related) change, since it requires a pass to pack the assignments into one bulk assignment. This change specifically avoids "unpacking" assignments in the first place (rather than relying on a pass to "repack" the assignments later on). I figured I would keep those two changes separate as this approach covers the more common case and avoids the complexity of expanding followed by packing. However that test demonstrates a reason why we may still want to try to "pack" assignments when the user code could be simplified.