[when] hotfix wiring to array with delayed elaboration (#1236)
Issue: the latest when bit-blasting changes made an assumption that the output resolve had already happened if it was going to happen, so it doesn’t work when the output (something assigned inside the when) is resolved/elaborated after the when logic is complete

Fix: need to add logic to handle elaborated when output, the essence of this is removing the guards in the when logic that allowed the child logic to only work on inputs, however this also means the output (hardware_module) needs to also assign the final output if it's elaborated. When doing this, I added concat logic that handles arrays that are combinations of single element children and slices. This works by using array create to combine neighboring children and concat to combine the resulting arrays including slices. This is opposed to the original fix which flattened slices using arrayget ops and a single array concat at the end.