[When] hotfix handling of wiring directly to when output (#1237)
Currently the when logic assumes that a when output only drives a single value, this invariant is broken when a user calls .value() on a conditionally driven value, getting a handle to a when output and wiring it as they'd wish.

This change preserves the invariant by:

Marking when output values with a flag which will trigger different wiring behavior
When directly wiring a when output to a new drivee, it behaves as if the new drivee was driven by same drivers as the current drivee. This is done by adding the conditional wires corresponding to the current drivee to the appropriate when contexts (essentially replaying the wiring of the current drivee, except with the new drivee).
In this process, I discovered a related issue in which the parent array logic assumes that the root array is assigned in the when. Instead, we may have an ancestor that is not the root (e.g. we could bulk assign elements of a 2d array). This updates the root logic to instead iterate over the ancestors towards the root until the appropriate value is found.