[when] Temporary Guard Against ALWCOMBORDER issue
See #1248 for context, adds a pass that checks whether this issue would
be possible and raises an exception.

Ideally we'd support the ok version of this which is when the variable
in the loop is assigned before it is used, but this requires some more
complex analysis which we can add incrementally (it's also a significant
step towards a more general solution that manages these dependencies).