[MLIR] Add locations to HW module signatures (#1278)
Due to https://github.com/llvm/circt/pull/4540, module ports now are also given
a location. We need to explicitly set their location (as we do for all op's) so
that they don't print `<stdin>...`.

It should be noted that in this PR, we only add the plumbing for module ports to
emit a location -- we don't actually emit valid locations, just the default
`loc(unknown)` for now. This brings us to parity before the
https://github.com/llvm/circt/pull/4540 change. Soon, we should actually emit
real locations of port declarations to enhance the annotations.