[When] Revert hotfix handling of wiring directly to when output (#1237)
This change duplicated when output targets when a user directly wires
the output of a when primitive.  This avoided an inconsistency that
occured when a when output was directly wired by the user and then
resolved after the direct wiring.  Since the reference to the direct
wiring wasn't stored, the direct wiring wasn't updated to match the
resolved value.  This was a problem when resolving values elaborated the
children as separate primitive outputs (rewiring the conditionally
driven value to new outputs).

Now that we are elaborating when outputs using the children of the
output (instead of creating new outputs for each child) we no longer
have the issue of needing to track additional values driven by a when
output (the standard resolve logic now handles this since we maintain
the origin when output as the driver).