[When] Remove invalid assertion
This assertion does not always hold True because `port.driving()`
could be empty, but represented as `[[], [], []]` if the value
is resolved.  We could improve the logic to check arbitrarily flattened
values, but I don't think it's worth the cost/complexity since this was
originally used as part of the debugging/development process rather than
useful for the end user.